# optimize implementation for large policies