```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkIsAllowed/SingleStatementAllow-8     413           427           +3.39%
BenchmarkIsAllowed/MultipleStatements-8       7884          677           -91.41%
BenchmarkIsAllowed/DenyRule-8                 851           832           -2.23%
BenchmarkIsAllowed/WildcardMatching-8         400           401           +0.12%
```